requirements.txt
streamlit==1.48.1
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
plotly>=5.22.0
statsmodels>=0.14.2
xgboost>=2.0.3
google-generativeai>=0.7.0
firebase-admin>=6.5.0
pyrebase4==4.8.0
requests>=2.31.0
requests-toolbelt>=0.7.1
google-api-core>=2.25.1
google-cloud-storage>=3.3.0
cachecontrol>=0.14.3
pycryptodome>=3.20.0
